Fractional length output implemented
