+++ src/scripts/hand_tracking.py
@@ -123,58 +123,3 @@ class handDetector():
         
         return self.side + " -> " + str(self.countFingers)
         
-        
-def videoCapture():
-    # Camera capture
-    cap         = cv.VideoCapture(0)
-    i           = 0
-    tracking    = handDetector(detectionCon=0.75, maxHands=2)
-    # Hand landmarks 
-    ids         = [4, 8, 12, 16, 20]
-
-    # Verify camera errors
-    if(cap.isOpened() == False):
-        print("Error openning the video")
-
-    while(cap.isOpened()):
-        
-        success, frame  = cap.read()
-
-        # Flip frame to correct predict
-        frame = cv.flip(frame,1)
-
-        # Hand's contour
-        contour         = tracking.findHands(frame)
-        pose            = tracking.findPosition(frame)
-        i              += 1
-
-        # Detection and decosntruction of List to String
-        tracking.handsLabel(pose, ids)
-        num = tracking.labelText()
-
-        if success:
-            font     = cv.FONT_HERSHEY_COMPLEX
-            left     = (50,50)
-            right    = (380, 50)
-
-            if tracking.countFingers > 0:
-                if tracking.label == 'Left':
-                    cv.putText(frame, num, left, font, 1, (0,0,255), 2)
-                else:
-                     cv.putText(frame, num, right, font, 1, (0,0,255), 2)
-
-            cv.imshow('Frame', frame)
-            key = cv.waitKey(1)
-
-            # Exit by user hand
-            """ if tracking.handFingers == "01100":
-                break  """
-
-            # Exit by user using keyboard
-            if key == ord('q'):
-                break
-        else:
-            break
-
-    cap.release()
-    cv.destroyAllWindows()
